dra-tran-recon-manual/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.9
asyncpg==0.28.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pandas==2.1.1
google-analytics-data==0.18.2
httpx==0.26.0
apscheduler==3.10.4
cryptography==41.0.7
pytz==2023.3.post1
